split backend_gcf into two parts, one which includes Gcf and one that does not.
